feat: refactor file structure
