+++ lua/cp/ui/window.lua
@@ -75,7 +75,6 @@ function M.restore_layout(state, tile_fn)
     local source_file
     if source_files ~= '' then
       local files = vim.split(source_files, '\n')
-      -- Prefer known extensions first, but accept any extension
       local known_extensions = vim.tbl_keys(constants.filetype_to_language)
       for _, file in ipairs(files) do
         local ext = vim.fn.fnamemodify(file, ':e')
